% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ranks.R
\name{betterEqual}
\alias{betterEqual}
\title{Treatment is in better or equal position}
\usage{
betterEqual(treatpos, small.values, leagueTable)
}
\arguments{
\item{treatpos}{a  list with the treatment and posistion}
}
\description{
Treatment is in better or equal position
}
