% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Define condition}
\usage{
condition(fn, ...)
}
\arguments{
\item{fn}{Character string specifiying type of condition.}

\item{\dots}{Function arguments.}
}
\value{
A list.
}
\description{
Define condition
}
\details{
Argument \code{fn} can be any of the following character strings
which can be abbreviated: "alwaysTRUE", "betterEqual",
"sameHierarchy", "specificPosition", "retainOrder", "biggerCIV"

TODO: add more details on various functions (especially, function
arguments)
}
