% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{\%AND\%}
\alias{\%AND\%}
\title{Combine selections with AND}
\usage{
A \%AND\% B
}
\arguments{
\item{selections}{as list. Each selection comprises of a list with the
function and its arguments}
}
\value{
data.tree
}
\description{
Combine selections with AND
}
\examples{
A <- condition("retainOrder", c("Placebo", "Salmeterol", "SFC"))
B <- condition("retainOrder", c("Placebo", "Salmeterol", "SFC"))
A \%AND\% B

}
