% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ranks.R
\name{isbiggerCIV}
\alias{isbiggerCIV}
\title{Treatment difference bigger than a given Clinically Important Value}
\usage{
isbiggerCIV(args, small.values, leagueTable)
}
\arguments{
\item{treatBigger}{the name of the treatment with the bigger absolute effect
irrespective of whether it is considered bad or good}

\item{civ}{(Clinicaly Important Value) is an optional argument 
of type list(treatBig, treatSmall, threshold) where
hierarchies fulfilling condition \code{treatBig - treatSmall > civ}  will
be recorded.}
}
\value{
True / False
}
\description{
Checks whether the effect of the first treatment is bigger than the second by
more than the given CIV.
The function does not take into account the \code{small.values=bad/good} 
argument.
}
