% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ranks.R
\name{isthesamerank}
\alias{isthesamerank}
\title{Ranks are equal}
\usage{
isthesamerank(ranklist, small.values, leagueTable)
}
\arguments{
\item{ranklist}{the rank input}

\item{rankProbabilityRow}{the rank from the \code{\link{precranking}} row output}
}
\value{
True / False
}
\description{
Checks whether two list of treatments are the same
}
