% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmarank.R
\name{nmaEffects}
\alias{nmaEffects}
\title{Checks whether the effects matrix \code{TE} and the
variance-covariance matrix have consistent names it}
\usage{
nmaEffects(TE, Cov)
}
\arguments{
\item{TE}{A matrix with the network estimates. Row and column names
should be present and correspond to the respective treatment
label. See example}

\item{Cov}{variance-covariance matrix of the relative treatment
effects row and column names should refer to the respective
comparison. Comparisons should be formatted as
\code{treat1:treat2}}
}
\value{
An object of class \code{nmaEffects} with \code{TE} a tibble with
the relative effects and \code{Cov} the variance-covariance matrix
provided as argument.
}
\description{
Checks whether the effects matrix \code{TE} and the
variance-covariance matrix have consistent names it
}
\examples{

data("Woods2010")
p1 <- pairwise(treatment, event = r, n = N,
               studlab = author, data = Woods2010, sm = "OR")
# Conduct network meta-analysis
net1 <- netmeta(p1)
TE <- net1$TE.random
Cov <- net1$Cov.random
nmaEffects(TE, Cov)

}
