% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{opposite}
\alias{opposite}
\title{The NOT function for a selection statement It simply reverses
condition}
\usage{
opposite(A)
}
\arguments{
\item{A}{selection list}
}
\description{
The NOT function for a selection statement It simply reverses
condition
}
\examples{
A = condition("retainOrder", c("Placebo", "Salmeterol", "SFC"))
opposite(A)

}
