% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ranks.R
\name{rankFromLeagueTable}
\alias{rankFromLeagueTable}
\title{Treatment hierarchy from smaller to bigger from the league table}
\usage{
rankFromLeagueTable(small.values, leagueTable)

rankFromLeagueTable(small.values, leagueTable)
}
\arguments{
\item{leagueTable}{is a table like the one exported by the 
\code{prepareNMAEffects} {$TE} object}
}
\value{
a vector of treatment names

a vector of treatment names
}
\description{
Treatment hierarchy from smaller to bigger from the league table

Treatment hierarchy from smaller to bigger from the league table
}
