% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ranks.R
\name{retainOrder}
\alias{retainOrder}
\title{Treatments in specified order}
\usage{
retainOrder(treatments, small.values, leagueTable)
}
\arguments{
\item{treatments}{a vector of treatments}

\item{rankProbabilityRow}{the rank from the \code{\link{precranking}} row output}
}
\value{
True / False
}
\description{
Checks whether the treatments specified as a vector are in the same order. 
Other treatments may be between them.
}
\examples{
 retainOrder(list("Placebo", "SFC", "salmeterol"), precisionsRow)

}
