% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmarank.R
\name{selectionHolds}
\alias{selectionHolds}
\title{Checks if a selection on relative treatment effects is true}
\usage{
selectionHolds(node, small.values, leagueTable)
}
\arguments{
\item{node}{of a predicate tree containing the name of a function}

\item{small.values}{"good" or "bad" defines rank order}

\item{leagueTable}{the \code{TE} of a \code{nmaEffects} object produced by the 
\code{prepareNMAEffects} function}
}
\value{
True/False
}
\description{
Calculates the value of a predicate statement predicate tree given 
a league table
}
