% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ranks.R
\name{treatementInSpecificPosition}
\alias{treatementInSpecificPosition}
\title{Treatment in position}
\usage{
treatementInSpecificPosition(treatpos, small.values, leagueTable)
}
\arguments{
\item{treatpos}{a  list with the treatment and posistion}

\item{rankProbabilityRow}{the rank from the \code{\link{precranking}} row output}
}
\value{
True / False
}
\description{
Checks whether a treatment is in specified position
}
