% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmarank.R
\name{nmaEffects}
\alias{nmaEffects}
\title{Check input data for \code{nmarank}}
\usage{
nmaEffects(TE, Cov)
}
\arguments{
\item{TE}{A matrix with the network meta-analysis relative treatment effects.
Rows and columns should be labeled according to the treatment names as
shown in the example.}

\item{Cov}{Variance-covariance matrix of the network meta-analysis relative
treatment effects. Row and column names should refer to the respective
treatment comparisons as shown in the example.}
}
\value{
An object of class \code{nmaEffects} with \code{TE} a tibble with
the relative effects and \code{Cov} the variance-covariance matrix
provided as argument.
}
\description{
Checks whether the network meta-analysis treatment effects TE and their
variance-covariance matrix Cov have consistent names.
}
\examples{

data("Woods2010")
p1 <- pairwise(treatment, event = r, n = N,
               studlab = author, data = Woods2010, sm = "OR")
# Conduct network meta-analysis
net1 <- netmeta(p1)
TE <- net1$TE.random
Cov <- net1$Cov.random
nmaEffects(TE, Cov)

}
