% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmarank.R
\name{nmarank}
\alias{nmarank}
\title{Precision of treatment ranking}
\usage{
nmarank(
  x,
  nsim = 10000,
  small.values = "bad",
  predicate = list(fn = "alwaysTrue", args = list())
)
}
\arguments{
\item{x}{either an object of class \code{\link{netmeta}} or a list with the 
effects and the variance-covariance list(TE=TE, Cov=Cov)}

\item{nsim}{number of simulations}

\item{small.values}{A character string specifying whether small treatment 
effects indicate a "good" or "bad" effect}

\item{predicate}{defines the condition that a possible relative effects 
vector should comply to. See \code{select-ranks}}
}
\value{
The main result is the \code{Output} is the list of rankings as lists 
of treatments and their probabilities
}
\description{
The main function of the package\code{\link{nmarank}} gives the probabilities 
of treatment hierarchies of network meta-analysis.
}
