% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmarank.R
\name{opposite}
\alias{opposite}
\title{The NOT function for a selection statement
It simply reverses predicate}
\usage{
opposite(A)
}
\arguments{
\item{A}{selection list}
}
\description{
The NOT function for a selection statement
It simply reverses predicate
}
\examples{
A = list(fn = "retainOrder", args = c("Placebo", "Salmeterol", "SFC"))
opposite(A)
 @export 
}
