% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{opposite}
\alias{opposite}
\title{The NOT function for a selection statement It simply reverses
condition}
\usage{
opposite(cond)
}
\arguments{
\item{cond}{\code{\link{condition}}.}
}
\value{
Object of class 'data.tree'.
}
\description{
The NOT function for a selection statement It simply reverses
condition
}
\examples{
data("Woods2010", package = "netmeta")
p1 <- pairwise(treatment, event = r, n = N, studlab = author,
  data = Woods2010, sm = "OR")
net1 <- netmeta(p1, small.values = "good")

A = condition("retainOrder", c("Placebo", "Salmeterol", "SFC"))
nmarank(net1, opposite(A), text.condition = "NOT order P-S-S", nsim = 5000)

}
\seealso{
\code{\link{condition}}, \code{\link{nmarank}}
}
